+++ src/plugin/parquet-writer-hl.cpp
@@ -64,7 +64,10 @@ ST_retcode sf_hl_write_varlist(
     // Get variable names
     // ------------------
 
-    if ( any_rc ) goto exit;
+    if ( any_rc ) {
+        rc = any_rc;
+        goto exit;
+    }
 
     j = 0;
     fstream.open(fcols);
@@ -151,7 +154,7 @@ ST_retcode sf_hl_write_varlist(
                 goto exit;
             }
         }
-        sf_running_timer (&timer, "Copied data from memory into Arrow table");
+        sf_running_timer (&timer, "Copied data into Arrow table");
 
         std::shared_ptr<arrow::Schema> schema = arrow::schema(vfields);
         std::shared_ptr<arrow::Table> table = arrow::Table::Make(schema, varrays);