+++ changelog.md
@@ -1,6 +1,16 @@
 Change Log
 ==========
 
+## parquet-0.4.0 (2018-10-31)
+
+### Features
+
+- Added column selector to file reader.
+- High-level file reader. Faster but might suffer from a loss of generality.
+- Scan string fields (ByteArray) to figure out the maximum length.
+  Optionally specify `nostrscan` to fall back to `strbuffer`; optionally
+  specify `strscan(.)` to scan the entire column.
+
 ## parquet-0.3.0 (2018-10-31)
 
 ### Features